% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{archive_copernicus}
\alias{archive_copernicus}
\title{Archive a stars object to a database}
\usage{
archive_copernicus(x, path = ".", ...)
}
\arguments{
\item{x}{stars object}

\item{path}{char, the data path}

\item{...}{arguments for \code{\link{generate_filename}}}
}
\value{
tabular database as a tibble
}
\description{
Archive a stars object to a database
}
