% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bind_stars}
\alias{bind_stars}
\title{Bind a list of stars objects into a multi-attribute stars object}
\usage{
bind_stars(x)
}
\arguments{
\item{x}{list of stars objects}
}
\value{
a single multi-attribute stars object
}
\description{
Bind a list of stars objects into a multi-attribute stars object
}
