% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{browse_product}
\alias{browse_product}
\title{Read a tabular catalog}
\usage{
browse_product(product = "GLOBAL_ANALYSISFORECAST_BGC_001_028")
}
\arguments{
\item{product}{chr, the product id}

\item{name}{Browse a product by id}
}
\value{
hmmm, whatever [httr](BROWSE) returns
}
\description{
Read a tabular catalog
}
