% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{build_cli_subset}
\alias{build_cli_subset}
\title{Build a CLI subset request}
\usage{
build_cli_subset(
  dataset_id = "cmems_mod_glo_phy-cur_anfc_0.083deg_P1D-m",
  product = "global-analysis-forecast-phy-001-024",
  vars = unlist(dplyr::pull(dplyr::filter(product_lut(product[1]), datasetid ==
    dataset_id), .data$variables)),
  bb = c(xmin = 5, ymin = 38, xmax = 10, ymax = 42),
  depth = c(0, 10),
  time = c("2022-01-01", "2022-01-15"),
  ofile = sprintf("\%s_\%s_\%s.nc", product, dataset_id, format(time[1], "\%Y-\%m-\%d")),
  extra = "--overwrite",
  app = get_copernicus_app(),
  log_level = "INFO"
)
}
\arguments{
\item{dataset_id}{char, the data set identifier such as 'cmems_mod_glo_phy-cur_anfc_0.083deg_P1D-m'}

\item{product}{chr the product family such as "global-analysis-forecast-phy-001-024"}

\item{vars}{char, a vector of one or more variables such as c("uo", "vo")}

\item{bb}{\code{\link[sf]{bbox}} or named numeric vector, either a \code{\link[sf]{bbox}} or 
a 4 element named vector with "xmin", "xmax", "ymin" and "ymax" named elements}

\item{depth}{numeric of NULL, two element vector of min and max depths}

\item{time}{POSIXct, Date or character, start and stop times}

\item{ofile}{char, the output filename (default is "./output.nc")}

\item{extra}{NULL or character, any other arguments for \code{copernicusmarine subset}}

\item{app}{char, the name of the application to run (default is "copernicusmarine")}
}
\value{
named 2 element character vector of the app and the args
copernicusmarine subset -i cmems_mod_glo_phy-cur_anfc_0.083deg_P1D-m -x 5.0 -X 10.0 -y 38.0 -Y 42.0 -z 0. -Z 10. -v uo -v vo -t 2022-01-01 -T 2022-01-15 -o ./copernicus-data -f dataset_subset.nc
}
\description{
See \href{https://help.marine.copernicus.eu/en/articles/7972861-copernicus-marine-toolbox-cli-subset#h_a906235d0a}{the docs}
}
