% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tile_copernicus.R
\name{check_tile_copernicus}
\alias{check_tile_copernicus}
\title{Check availability of a COPERNICUS tile}
\usage{
check_tile_copernicus(tileH,tileV)
}
\arguments{
\item{tileH}{H index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{tileV}{V index of the tile COPERNICUS system (\code{numeric} or \code{character})}
}
\value{
a logical \code{vector} indicating which pair (H,V) is available
}
\description{
Check whether the provided (H,V) pairs are available as COPERNICUS tiles
}
\details{
\code{tileH} and \code{tileV} should have the same length and form pairs of (H,V) indices. Each tile corresponds to a file covering an area of 10 x 10
}
\examples{
check_tile_copernicus(19,4) # H19V4 is in Italy
check_tile_copernicus(0,4) # H0V4 is not available

}
\author{
Antoine Stevens
}
\references{
Baret et al. 2010. BioPar Product User Manual. Geoland2: Towards an Operational GMES Land Monitoring Core Service, 42 p.
}

