% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_time_copernicus.R
\name{check_time_copernicus}
\alias{check_time_copernicus}
\title{Begin and end date of COPERNICUS products}
\usage{
check_time_copernicus(product,begin,end)
}
\arguments{
\item{product}{one of the following: 'NDVI_V1' (Normalized Difference Vegetation Index - VGT instrument),'NDVI_V2' (Normalized Difference Vegetation Index - PROBAV instrument),'LAI' (Leaf Area Index),'FCOVER' (Fraction of Vegetation Green Cover),
'FAPAR' (Fraction of Absorbed Photosynthetically Active Radiation),'VCI' (Vegetation Condition Index),'VPI' (Vegetation Productivity Index),
'DMP' (Dry Matter Productivity),'BA' (Burnt Areas)}

\item{begin}{begin of the time period. \code{Date} object, \code{numeric} or \code{character} of length 1 that can be transformed to a \code{Date} using \code{\link[lubridate]{ymd}}. See \code{?lubridate::ymd} for more details. \code{ymj} format is also accepted (with \code{j} being the day of the year)}

\item{end}{end of the time period. Same format as \code{begin}}
}
\value{
a \code{list} with components \code{begin} and \code{end} strip to the available time period
}
\description{
Check whether a COPERNICUS product is availabe within a given time period
}
\examples{
# Check data availability for NDVI V1, between 01-01-2010 and 31-12-2014
check_time_copernicus('NDVI_V1',20100101,20141231)
# Check data availability for NDVI V2, between 01-01-2010 and 31-12-2014
# using different date formats
check_time_copernicus('NDVI_V2','2010001','2014365')
check_time_copernicus('NDVI_V2','2010-01-01','2014-12-31')

}
\author{
Antoine Stevens
}

