% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_version_copernicus.R
\name{check_version_copernicus}
\alias{check_version_copernicus}
\title{Check algorithm version for a given COPERNICUS product}
\usage{
check_version_copernicus(product,server,user,password)
}
\arguments{
\item{product}{one of the following: 'NDVI_V1' (Normalized Difference Vegetation Index - VGT instrument),'NDVI_V2' (Normalized Difference Vegetation Index - PROBAV instrument),'LAI' (Leaf Area Index),'FCOVER' (Fraction of Vegetation Green Cover),
'FAPAR' (Fraction of Absorbed Photosynthetically Active Radiation),'VCI' (Vegetation Condition Index),'VPI' (Vegetation Productivity Index),
'DMP' (Dry Matter Productivity),'BA' (Burnt Areas)}

\item{server}{url of the COPERNICUS data server. Default to 'http://land.copernicus.vgt.vito.be/PDF///datapool/Vegetation/', see 'copernicus_options('server')'}

\item{user}{user name to access COPERNICUS data portal. Default set via \code{copernicus_options('user')}}

\item{password}{password associated with user name to access COPERNICUS data portal. Default set via \code{copernicus_options('password')}}
}
\value{
a \code{character} vector of algorithm version
}
\description{
Retrieve version of the algorithm before and after May/June 2014
}
\details{
Versions of the processing algorithm may be subject to change.
This function fetches the server data portal to retrieve up-to-date algorithm versions for each product,
taking into account changes in product version after May/June 2014
}
\examples{
\dontrun{
# user and password should be provided, for instance through global package options
# ?copernicus_options
# NDVI_V1: there is only one version of the algorithm
check_version_copernicus('NDVI_V1')
#  LAI
check_version_copernicus('LAI')
}
}
\author{
Antoine Stevens
}

