% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{compose_filename_v0.01}
\alias{compose_filename_v0.01}
\title{Compose a filename from a database}
\usage{
compose_filename_v0.01(x, path = ".", ext = ".tif")
}
\arguments{
\item{x}{database (tibble), with date, var, depth}

\item{path}{character, the root path for the filename}

\item{ext}{character, the filename extension to apply (with dot)}
}
\value{
character vector of filenames in form
        \code{<path>/YYYY/mmdd/YYYY-mm-dd_var_depth.tif}
}
\description{
Compose a filename from a database
}
