% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus-deprecated.R, R/path.R
\name{copernicus-deprecated}
\alias{copernicus-deprecated}
\alias{get_credentials}
\alias{set_credentials}
\title{Deprecated functions in package \pkg{copernicus}.}
\usage{
get_credentials(filename = "~/.copernicuscredentials", parse = TRUE)

set_credentials(x = "username:password", filename = "~/.copernicuscredentials")
}
\arguments{
\item{filename}{the name the file to store the credentials as a single line of text}

\item{parse}{logical, if TRUE split the credentials and return as a named
character vector}

\item{x}{char, credentials as \code{"username:password"}}
}
\value{
character \code{username:password}, a named vector or or NULL if the file is not found.

NULL invisibly
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\keyword{internal}
