% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus-package.R
\docType{package}
\name{copernicus-package}
\alias{copernicus-package}
\alias{copernicus}
\title{copernicus: Serves Center for Ocean Forecasting Copernicus needs}
\description{
Provides download, archiving and access to Copernicus related local datasets.

Tools for working with Copernicus data.
}
\keyword{internal}
