% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{copernicus_path}
\alias{copernicus_path}
\title{Retrieve the copernicus path}
\usage{
copernicus_path(..., root = "/mnt/ecocast/coredata/copernicus")
}
\arguments{
\item{...}{further arguments for \code{file.path()}}

\item{root}{the root path}
}
\value{
character path description
}
\description{
Retrieve the copernicus path
}
