% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{create_dataset_metadata}
\alias{create_dataset_metadata}
\title{Create the table of dataset metadata}
\usage{
create_dataset_metadata(x = read_product_catalog())
}
\arguments{
\item{x}{catalog (flattened or unflattened)}
}
\value{
table narrowed and depduplictaed
}
\description{
Create the table of dataset metadata
}
