% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{create_lut}
\alias{create_lut}
\title{Given a NCDF object (or path to one) create a variable lookup table (LUT)}
\usage{
create_lut(x)
}
\arguments{
\item{x}{ncdf4 object or path to one, if the latter we will open and then close it}
}
\value{
tibble LUT of \code{name}, \code{longname} and \code{units}
}
\description{
Given a NCDF object (or path to one) create a variable lookup table (LUT)
}
