% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{database_variables}
\alias{database_variables}
\title{Retrieve the database pre-defined variable names}
\usage{
database_variables()
}
\value{
charcater vector of variable names
}
\description{
Retrieve the database pre-defined variable names
}
