% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{dataset_is_climatology}
\alias{dataset_is_climatology}
\title{Guess if a dataset is a climatology dataset}
\usage{
dataset_is_climatology(
  x = c("cmems_mod_glo_phy_anfc_0.083deg_P1D-m", "cmems_mod_glo_phy_anfc_0.083deg_P1M-m",
    "cmems_mod_glo_phy_my_0.083deg-climatology_P1M-m",
    "cmems_mod_glo_phy_my_0.083deg_static")
)
}
\arguments{
\item{x}{one or more dataset_id values}
}
\value{
one logical per element of input `x`, TRUE if climatology
}
\description{
Guess if a dataset is a climatology dataset
}
