% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename}
\alias{decompose_filename}
\title{Decompose a filename into a database}
\usage{
decompose_filename(x = "2021-03-20_zos_sur.tif", ext = ".tif")
}
\arguments{
\item{x}{character, vector of one or more filenames}

\item{ext}{character, the extension to remove (including dot)}
}
\value{
table (tibble) database
}
\description{
Decompose a filename into a database
}
