% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename_pre_v2}
\alias{decompose_filename_pre_v2}
\title{Decompose a filename into a database}
\usage{
decompose_filename_pre_v2(
  x = c("2025-03-17_day_uo_sur_none.tif", "2019-01-01_day_mlotst_sur_none.tif"),
  ext = ".tif"
)
}
\arguments{
\item{x}{character, vector of one or more filenames}

\item{ext}{character, the extension to remove (including dot)}
}
\value{
table (tibble) database
}
\description{
Decompose a filename into a database
}
