% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_copernicus.R
\name{download_copernicus}
\alias{download_copernicus}
\title{Download COPERNICUS products}
\usage{
download_copernicus(product,begin,end,extent,tileH,tileV,
                    outPath,user,password,allowParallel = FALSE,...)
}
\arguments{
\item{product}{One of the following: 'NDVI_V1' (Normalized Difference Vegetation Index - VGT instrument),'NDVI_V2' (Normalized Difference Vegetation Index - PROBAV instrument),'LAI' (Leaf Area Index),'FCOVER' (Fraction of Vegetation Green Cover),
'FAPAR' (Fraction of Absorbed Photosynthetically Active Radiation),'VCI' (Vegetation Condition Index),'VPI' (Vegetation Productivity Index),
'DMP' (Dry Matter Productivity),'BA' (Burnt Areas)}

\item{begin}{begin of the time period. \code{Date} object, \code{numeric} or \code{character} of length 1 that can be transformed to a \code{Date} using \code{\link[lubridate]{ymd}}. See \code{?ymd} for more details. \code{ymj} format is also accepted (with \code{j} being the day of the year)}

\item{end}{end of the time period. Same format as \code{begin}}

\item{extent}{an object of class \code{Raster*}, \code{Extent} or \code{SpatialPolygons*} giving the extent from which tiles should be downloaded.
See \code{\link{get_tile_copernicus}} for details.}

\item{tileH}{H index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{tileV}{V index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{outPath}{Path where downloaded files should be stored. Default set via \code{copernicus_options('downloadPath')}}

\item{user}{User name to access COPERNICUS data portal. Default set via \code{copernicus_options('user')}}

\item{password}{Password associated with user name to access COPERNICUS data portal. Default set via \code{copernicus_options('password')}}

\item{ntry}{Number of tries after non-succes download response from the server. Increasing the value increases chances to actually get the data. Default set via \code{copernicus_options('ntry')}}

\item{allowParallel}{Logical. If a \code{foreach} parallel backend is loaded and available, should the function use it? Default is \code{FALSE}.}

\item{...}{argument passed to \code{\link{get_url_copernicus}}, such as groupByDate}
}
\value{
Return invisibly the list of downloaded files
}
\description{
Download COPERNICUS products, for a given period of time and tile(s)
}
\details{
If target files are already present in the \code{outPath} directory, they are not downloaded.
One should choose to use either an extent or pairs of (H,V) values. If extent is provided, tileH and tileV are not used.
}
\examples{
\dontrun{
# Don't forget to provide in copernicus_options() your user and password
# for COPERNICUS data portal before running this
# Let's download NDVI V1 prodct, for JUN 2009, and tile (h=19,v=4)
download_copernicus(product = 'NDVI_V1', begin = '2009-06-01', end = '2009-06-31',
                   tileH = 19, tileV = 4)
# one could also use an extent object, instead of (tileH,tileV) pairs, eg:
library(raster)
e <- extent(c(-1,2,49,51))
# this will download 12 files (4 tiles x 3 time periods)
download_copernicus(product = 'NDVI_V1', begin = '2009-06-01', end = '2009-06-31', extent = e)
}

}
\author{
Antoine Stevens
}

