% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{download_copernicus_cli}
\alias{download_copernicus_cli}
\title{Fetch via copernicus-marine subset}
\usage{
download_copernicus_cli(..., verbose = FALSE, credentials = get_credentials())
}
\arguments{
\item{...}{arguments for \code{\link{build_cli_subset}}}

\item{verbose}{logical, if true pint the calling sequence excluding credentials}

\item{credentials}{two element named character vector of \code{username} and \code{password}}
}
\value{
numeric, 0 for success
}
\description{
Fetch via copernicus-marine subset
}
