% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{fetch_cli_subset}
\alias{fetch_cli_subset}
\title{Fetch via copernicus-marine subset}
\usage{
fetch_cli_subset(..., verbose = FALSE, credentials = get_credentials())
}
\arguments{
\item{...}{arguments for \code{\link{build_cli_subset}}}

\item{verbose}{logical, if true pint the calling sequence excluding credentials}

\item{credentials}{two element named character vector of \code{username} and \code{password}}
}
\value{
numeric, 0 for success
}
\description{
Fetch via copernicus-marine subset
}
