% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{fetch_copernicus_cli_subset}
\alias{fetch_copernicus_cli_subset}
\title{Fetch Copernicus data as a list of \code{stars} objects}
\usage{
fetch_copernicus_cli_subset(
  ofile = copernicus_path("temp", paste0(dataset_id[1], ".nc")),
  cleanup = TRUE,
  form = c("stars", "filename", "list")[1],
  group = TRUE,
  ...
)
}
\arguments{
\item{cleanup}{logical, if TRUE clean up files}

\item{form}{str, one of "stars" or "filename" to return a stars object or the filename
If "filename" then no attempt is made to read the downloaded file and `cleanup` is ignored.
If "list" then read the file into a list of stars objects where variables are grouped
by dimensional dependence.}

\item{group}{logical, if `form` is "list" and if TRUE read variables grouped by dimensional dependence
if FALSE then each variable is read separately}

\item{...}{
  Arguments passed on to \code{\link[=download_copernicus_cli_subset]{download_copernicus_cli_subset}}
  \describe{
    \item{\code{verbose}}{logical, if true pint the calling sequence excluding credentials}
    \item{\code{tries}}{numeric, the number of attempts permitted if a fail occurs}
  }}
}
\value{
stars object or NULL
}
\description{
This is a wrapper around \code{\link{download_copernicus_cli_subset}} that
hides the details and returns a \code{stars} object. Be aware that what gets
returned may have one or more degenerate dimensions (single element dimensions)
such as a single depth, time or both; it depends upon the request that you make.
}
