% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{fetch_copernicus_script}
\alias{fetch_copernicus_script}
\title{Fetch Copernicus data as a list of \code{stars} objects}
\usage{
fetch_copernicus_script(
  script = "cmems_mod_glo_phy",
  date = Sys.Date(),
  out_path = tempfile(pattern = "copernicus", tmpdir = tempdir(), fileext = ".nc"),
  cleanup = TRUE,
  ...
)
}
\arguments{
\item{script}{character, the name of the script to use}

\item{date}{Date class or castable as Date}

\item{out_path}{character, the temporary path to store the downloaded file}

\item{cleanup}{logical, if TRUE clean up files}

\item{...}{further arguments for \code{\link{populate_script}}}
}
\value{
named list of stars objects (organized by variable)
}
\description{
This is a wrapper around \code{\link{download_copernicus_script}} that
hides the details and returns a list of \code{stars} objects.  The downloaded
file is deleted.
}
