% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{fetch_dataset_catalog}
\alias{fetch_dataset_catalog}
\title{Fetch catalog}
\usage{
fetch_dataset_catalog(
  filename = copernicus_path("catalogs/all_products_and_datasets.json"),
  app = get_copernicus_app()
)
}
\arguments{
\item{filename}{chr the name of the file to download}

\item{app}{chr the name of the app to run}
}
\value{
catalog list
}
\description{
Fetch catalog
}
