% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{fetch_dataset_description}
\alias{fetch_dataset_description}
\title{Fetch the description for a dataset}
\usage{
fetch_dataset_description(
  dataset_id = "cmems_mod_glo_phy_myint_0.083deg_P1D-m",
  app = get_copernicus_app(),
  ofile = copernicus_path("catalogs", sprintf("\%s.json", dataset_id[1]))
)
}
\arguments{
\item{dataset_id}{chr, the dataset to describe}

\item{app}{chr, the `copernicusmarine` app}

\item{ofile}{chr, the name of the file to save with thr response content}
}
\value{
0 for success non-zero otherwise
}
\description{
Fetch the description for a dataset
}
