% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_product_by_day}
\alias{fetch_product_by_day}
\title{A convenience tool for fetching a product suite on a given day.}
\usage{
fetch_product_by_day(
  p = dplyr::filter(product_lut(), .data$fetch == "yes", .data$product_id ==
    .data$product_id[1]),
  x = Sys.Date() - 7,
  bb = sf::st_bbox(c(xmin = -180, ymin = -90, xmax = 180, ymax = 90), crs = 4326),
  ...
)
}
\arguments{
\item{p}{table of product info to be grouped by `dataset_id`}

\item{x}{Date (or YYYY-mm-dd string)}

\item{bb}{bounding box or somehting from which a bounding box can be found.}

\item{...}{other arguments for [fetch_copernicus_cli_subset]}
}
\value{
list of stars objects (one for each dataset) possibly with NULLs for unknown datasets
}
\description{
This only works for known products/datasets.  Unknowns will generate a
warning and return NULL. For each dataset only a single depth layer is 
returned.
}
