% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{fetch_product_catalog}
\alias{fetch_product_catalog}
\title{Fetch the product catalog}
\usage{
fetch_product_catalog(
  product_id = "GLOBAL_ANALYSISFORECAST_PHY_001_024",
  app = get_copernicus_app(),
  ofile = copernicus_path("catalogs", sprintf("\%s.json", product_id[1]))
)
}
\arguments{
\item{product_id}{chr, the product to describe or "all_products" to fetch all}

\item{app}{chr, the `copernicusmarine` app}

\item{ofile}{chr, the name of the file to save with the response content}
}
\value{
0 for success non-zero otherwise
}
\description{
Fetch the product catalog
}
