% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{flatten_product}
\alias{flatten_product}
\title{Flatten a particular product suite into one large table}
\usage{
flatten_product(
  x = dplyr::filter(read_product_catalog(), .data$product_id ==
    "GLOBAL_ANALYSISFORECAST_PHY_001_024", grepl("P1D", .data$dataset_id, fixed = TRUE))
)
}
\arguments{
\item{x}{nested catalog table}
}
\value{
a flattened tables
}
\description{
Flatten a particular product suite into one large table
}
