% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{flatten_variables}
\alias{flatten_variables}
\title{Flatten a dataset `variables` element}
\usage{
flatten_variables(x, dataset_id = "unknown")
}
\arguments{
\item{x}{a `variables` node}
}
\value{
table or possibly filled with NAs if x is NULL
}
\description{
Flatten a dataset `variables` element
}
