% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_tile_copernicus.R
\name{gen_tile_copernicus}
\alias{gen_tile_copernicus}
\title{Generate COPERNICUS tiling grid}
\usage{
gen_tile_copernicus(poly,offset)
}
\arguments{
\item{poly}{logical value indicating whether the function should return a . Default set to \code{FALSE}}

\item{offset}{numeric. Shift the tiling grid in the upper-left direction. Default to (1/112)/2.
The reference position is the centre of the pixel in COPERNICUS products, while the \code{raster} package, as well as other
GIS (eg QGIS) expect the reference of a grid to be at the upper-left corner. It means that the upper-left corner of the pixels is at  [(pixel_longitude - angular_resolution/2),(pixel_latitude + angular_resolution/2)]}
}
\value{
a \code{data.frame} or \code{SpatialPolygonsDataFrame}. See \code{poly} parameter
}
\description{
Generates the COPERNICUS tiling system, excluding tiles where there are no data available
}
\details{
See also \code{\link[MODIS]{genTile}} in the MODIS package
}
\examples{
# Return a data.frame with tiling info
df <- gen_tile_copernicus()
head(df)
# Return a Spatial object
library(sp)
pol <- gen_tile_copernicus(poly = TRUE)
plot(pol)
}
\author{
Antoine Stevens
}
\references{
Baret et al. 2010. BioPar Product User Manual. Geoland2: Towards an Operational GMES Land Monitoring Core Service, 42 p.
}

