% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_bounds}
\alias{get_bounds}
\title{Retrieve the native bounding box of the data}
\usage{
get_bounds(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
4 element vector of \code{[xmin, xmax, ymin, ymax]}
}
\description{
Retrieve the native bounding box of the data
}
