% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_copernicus.R
\name{get_copernicus}
\alias{get_copernicus}
\title{Download COPERNICUS products}
\usage{
get_copernicus(product,begin,end,tileH,tileV,outPath,user,password)
}
\arguments{
\item{product}{One of the following: 'NDVI_V1' (Normalized Difference Vegetation Index - VGT instrument),'NDVI_V2' (Normalized Difference Vegetation Index - PROBAV instrument),'LAI' (Leaf Area Index),'FCOVER' (Fraction of Vegetation Green Cover),
'FAPAR' (Fraction of Absorbed Photosynthetically Active Radiation),'VCI' (Vegetation Condition Index),'VPI' (Vegetation Productivity Index),
'DMP' (Dry Matter Productivity),'BA' (Burnt Areas)}

\item{begin}{begin of the time period. \code{Date} object, \code{numeric} or \code{character} of length 1 that can be transformed to a \code{Date} using \code{\link[lubridate]{ymd}}. See \code{?ymd} for more details. \code{ymj} format is also accepted (with \code{j} being the day of the year)}

\item{end}{end of the time period. Same format as \code{begin}}

\item{tileH}{H index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{tileV}{V index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{outPath}{Path where downloaded files should be stored. Default set via \code{copernicus_options('downloadPath')}}

\item{user}{user name to access COPERNICUS data portal. Default set via \code{copernicus_options('user')}}

\item{password}{password associated with user name to access COPERNICUS data portal. Default set via \code{copernicus_options('password')}}
}
\value{
Return invisibly the list of downloaded files
}
\description{
Download COPERNICUS products, for a given period of time and tile(s)
}
\details{
If target files are already present in the \code{outPath} directory, they are not downloaded
}
\examples{
\dontrun{
# Don't forget to provide in copernicus_options() your user and password
# for COPERNICUS data portal before running this
# this will download a lot of data!
get_copernicus(product = 'NDVI_V1', begin = '2009-06-01', end = '2014-05-31', tileH = 19, tileV = 4)
}
}
\author{
Antoine Stevens
}

