% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_copernicus_app}
\alias{get_copernicus_app}
\title{Retrieve `copernicusmarine` app}
\usage{
get_copernicus_app(x = "copernicusmarine", filename = "~/.copernicusapp")
}
\arguments{
\item{filename}{the name the file}
}
\value{
chr the known app (with path)
}
\description{
Retrieve `copernicusmarine` app
}
