% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_credentials}
\alias{get_credentials}
\title{Retrieve copernicus credentials.  One only needs credentials to download
data from the CMEMS servers.  Credentials are not required to access the locally
stored copernicus data (but file permissions are).}
\usage{
get_credentials(filename = "~/.copernicuscredentials", parse = TRUE)
}
\arguments{
\item{filename}{the name file that stores the credential in a one-line text file
as \code{username:password}}

\item{parse}{logical, if TRUE split the credentials and return as a named
character vector}
}
\value{
character \code{username:password}, a named vector or or NULL if the file is not found.
}
\description{
Retrieve copernicus credentials.  One only needs credentials to download
data from the CMEMS servers.  Credentials are not required to access the locally
stored copernicus data (but file permissions are).
}
