% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_dataset_node}
\alias{get_dataset_node}
\title{Get one dataset catalog node}
\usage{
get_dataset_node(x = read_dataset_catalog(), y = 1)
}
\arguments{
\item{x}{dataset catalog list}

\item{y}{number or name of the node to index}
}
\value{
a datasetnode (list with id, name and versions elements)
}
\description{
Get one dataset catalog node
}
