% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_dim}
\alias{get_dim}
\title{Retrieve dimension values by name}
\usage{
get_dim(x, what = c("time", "latitude", "longitude", "depth", "lon", "lat")[1])
}
\arguments{
\item{x}{ncdf4 object}

\item{what}{character, one of \code{time, latitude, longitude, depth, lon, lat}}
}
\value{
vector of numeric or POSIXct time
}
\description{
Retrieve dimension values by name
}
