% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_lon}
\alias{get_lon}
\alias{get_lat}
\alias{get_depth}
\alias{get_loc}
\title{Retrieve geospatial information about a grads resource}
\usage{
get_lon(x)

get_lat(x)

get_depth(x)

get_loc(x, what = "time")
}
\arguments{
\item{x}{ncdf4 object}

\item{what}{character, one of 'time', 'lon', 'lat', 'depth' or 'all'}
}
\value{
numeric longitude locations

numeric latitude locations

numeric depth locations

varies by value of what
\itemize{
  \item{\code{lon, lat, depth} numeric vectors}
  \item{\code{time} POSIXct vector}
  \item{\code{all} list of \code{lon, lat, depth, time} vectors}
}
}
\description{
Retrieve geospatial information about a grads resource
}
\section{Functions}{
\itemize{
\item \code{get_lon()}: retrieve numeric longitudes

\item \code{get_lat()}: retrieve numeric latitudes

\item \code{get_depth()}: retrieve numeric depths

}}
