% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_parts_value}
\alias{get_parts_value}
\title{Given a variables node, get a standard_name, short_name or units}
\usage{
get_parts_value(x, name = "standard_name")
}
\arguments{
\item{x}{a variables node with multiple parts}

\item{name}{chr, the desired element name}
}
\value{
character, possibly NA
}
\description{
Given a variables node, get a standard_name, short_name or units
}
