% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_range}
\alias{get_range}
\title{Get the native bounds for each dimension}
\usage{
get_range(x, what = "lon")
}
\arguments{
\item{x}{ncdf4 object}

\item{what}{character, one of 'time', 'lon', 'lat', 'depth' or 'all'}
}
\value{
varies by value of what
\itemize{
  \item{\code{lon, lat, depth} numeric vectors}
  \item{\code{time} POSIXct vector}
  \item{\code{all} list of the above}
}
}
\description{
Get the native bounds for each dimension
}
