% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile_copernicus.R
\name{get_tile_copernicus}
\alias{get_tile_copernicus}
\title{Get COPERNICUS tile(s) for a given extent}
\usage{
get_tile_copernicus(extent,tileH,tileV,...)
}
\arguments{
\item{extent}{an object of class \code{Raster*}, \code{Extent} or \code{SpatialPolygons*} giving the extent from which tiles info is retrieved}

\item{tileH}{H index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{tileV}{V index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{...}{argument passed to \code{\link{gen_tile_copernicus}}}
}
\value{
\code{data.frame} with tile(s) info: h, v, xmin, xmax, ymin, ymax
}
\description{
Get COPERNICUS tile(s) position and name for a given extent or tile (H,V) pairs
}
\details{
One should choose to use either an extent or pairs of (H,V) values
If \code{extent} is of class \code{Extent}, then it is supposed to be in geographical coordinates.
If \code{extent} is of class \code{Raster*} or \code{SpatialPolygons*} they should have a coordinate system.
Part of the code has been grabed from the \code{MODIS} package v0.31
}
\examples{
library(raster)
# using an extent object
e <- extent(10,25,45,50) # xmin,xmax,ymin,ymax
suppressWarnings(get_tile_copernicus(extent = e))
# using (H,V) pairs
get_tile_copernicus(tileH = 19:20, tileV = 4:5)
}
\author{
Antoine Stevens and Matteo Mattiuzzi (MODIS package)
}

