% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_url_copernicus.R
\name{get_url_copernicus}
\alias{get_url_copernicus}
\title{Get url's of COPERNICUS products}
\usage{
get_url_copernicus(product,begin,end,tileH,tileV,groupByDate,server,check_version = FALSE)
}
\arguments{
\item{product}{one of the following: 'NDVI_V1' (Normalized Difference Vegetation Index - VGT instrument),'NDVI_V2' (Normalized Difference Vegetation Index - PROBAV instrument),'LAI' (Leaf Area Index),'FCOVER' (Fraction of Vegetation Green Cover),
'FAPAR' (Fraction of Absorbed Photosynthetically Active Radiation),'VCI' (Vegetation Condition Index),'VPI' (Vegetation Productivity Index),
'DMP' (Dry Matter Productivity),'BA' (Burnt Areas)}

\item{begin}{begin of the time serie. \code{Date} object, \code{numeric} or \code{character} of length 1 that can be transformed to a \code{Date} using \code{\link[lubridate]{ymd}}. See \code{?ymd} for more details. \code{ymj} format is also accepted (with \code{j} being the day of the year)}

\item{end}{end of the time serie. Same format as \code{begin}}

\item{tileH}{H index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{tileV}{V index of the tile COPERNICUS system (\code{numeric} or \code{character})}

\item{groupByDate}{logical value indicating whether to return a list of vector with url's grouped by date, or a vector of url's. Default is \code{FALSE}.}

\item{server}{url of the COPERNICUS data server. Default to 'http://land.copernicus.vgt.vito.be/PDF///datapool/Vegetation/', see 'copernicus_options('server')'}

\item{check_version}{logical value indicating whether the algorithm version should be checked on the server data portal. Default to FALSE. See \code{\link{check_version_copernicus}}}
}
\value{
a character \code{vector} of url('s)
}
\description{
Get COPERNICUS url's for a given product, tile, and period of time
}
\examples{
# URL's for LAI product, for JAN 2014 and tiles (h=19,v=3) and (h=20,v=4)
get_url_copernicus(product = 'LAI',begin = '20140101', end = '20140131',tileH = 19:20, tileV = 3:4)
get_url_copernicus(product = 'NDVI_V2',begin = '20130101', end = '20130131',tileH = 19:20, tileV = 3:4)
}
\author{
Antoine Stevens
}

