% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_var_array}
\alias{get_var_array}
\title{Retrieve an array of data from a ncdf object}
\usage{
get_var_array(x, var, index, collapse_degen = FALSE)
}
\arguments{
\item{x}{ncdf4 object}

\item{var}{variable id (name)}

\item{index}{named list of \code{[start,length]} vectors for each dimension
\itemize{
\item{lon}{such as c(1,913)}
\item{lat}{such as c(1,443)}
\item{time}{such as c(5, 7) which starts on the 5th and ends on the 12th}
\item{depth}{such as c(8,1) which retrieves just the 8th depth}
}}

\item{collapse_degen}{logical, if FALSE then preserve length 1 dimensions
NAM218 grib files have variables with either \code{[lon, lat, time]} or
\code{[lon, lat, depth, time]} dimensions. If the grib files stored those in
\code{[lon, lat, time, depth]} order we could have dropped degenerate dimensions.}
}
\value{
matrix or array
}
\description{
Retrieve an array of data from a ncdf object
}
