% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_vardims}
\alias{get_vardims}
\title{Retrieve the dimension names for a variable}
\usage{
get_vardims(x, var = get_varnames(x)[1])
}
\arguments{
\item{x}{ncdf4 object}

\item{var}{character, the name of the variable}
}
\value{
character vector of variable names
}
\description{
Retrieve the dimension names for a variable
}
