% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_varnames}
\alias{get_varnames}
\alias{get_dimnames}
\alias{get_dims}
\alias{list_dims}
\title{Retrieve information about a grads resource}
\usage{
get_varnames(x, longnames = FALSE)

get_dimnames(x, longnames = FALSE)

get_dims(x)

list_dims(x)
}
\arguments{
\item{x}{ncdf4 object}

\item{longnames}{logical, if TRUE then returve the longnames}
}
\value{
character vector

named numeric vector of dimension lengths

named list of character vectors of dimension names per variable
}
\description{
Retrieve information about a grads resource
}
\section{Functions}{
\itemize{
\item \code{get_dimnames()}: Retrieve the names of available dimensions

\item \code{get_dims()}: Retrieve the dimensions

\item \code{list_dims()}: Retrieve the dimensions

}}
