% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{import_product_catalog}
\alias{import_product_catalog}
\title{Import the json product catalog}
\usage{
import_product_catalog(
  name = "all_products_copernicus_marine_service",
  path = copernicus_path("catalogs"),
  tabulate = TRUE,
  flatten = TRUE
)
}
\arguments{
\item{name}{chr, the name of the catalog name}

\item{path}{chr the path to the file}

\item{tabulate}{logical, if TRUE transform to a nested table}

\item{flatten}{logical, if TRUE and `tabulate` is TRUE then transform to a 
flat table}
}
\value{
a named list of json element, nested list or a table
}
\description{
Import the json product catalog
}
