% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_databases}
\alias{list_databases}
\title{Retrieve a list of databases}
\usage{
list_databases(path = copernicus_path(), pattern = "^database$")
}
\arguments{
\item{path}{chr, the root data directory}

\item{pattern}{chr, database filename regex pattern to search for}
}
\value{
database paths relative to the root path
}
\description{
Retrieve a list of databases
}
