% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{list_vardims}
\alias{list_vardims}
\title{Retrieve a table of variable dimensional dependencies}
\usage{
list_vardims(filename)
}
\arguments{
\item{filename}{str, the ncdf file to tabulate}
}
\value{
date frame of variables and dimensional dependencies
}
\description{
Retrieve a table of variable dimensional dependencies
}
