% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_path}
\alias{make_path}
\title{Make a path}
\usage{
make_path(x, recursive = TRUE, ...)
}
\arguments{
\item{x}{character path specification}

\item{recursive}{logical}

\item{...}{other argum,ents for \code{dir.create}}
}
\value{
logical with TRUE for success
}
\description{
Make a path
}
