% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{missing_records}
\alias{missing_records}
\title{Given a database (days, 8DR or month), determine the times
that are missing between the first and last records}
\usage{
missing_records(
  x,
  by = dplyr::pull(dplyr::slice(x, 1), dplyr::all_of("period"))
)
}
\arguments{
\item{x}{a database - typically filtered to just a single period}

\item{by}{chr, unit of time used for creating sequence to match against}
}
\value{
Date dates that seem to be missing
}
\description{
Given a database (days, 8DR or month), determine the times
that are missing between the first and last records
}
