% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{open_nc}
\alias{open_nc}
\title{Open a copernicus connection given a product id}
\usage{
open_nc(
  product_id = "cmems_mod_glo_phy-cur_anfc_0.083deg_P1D-m",
  credentials = get_credentials(),
  base_uri = "https://nrt.cmems-du.eu/thredds/dodsC"
)
}
\arguments{
\item{product_id}{char, the product identifer}

\item{credentials}{NULL or character vector}

\item{base_uri}{char base UIR for opendap}
}
\value{
ncdf4 object
}
\description{
Open a copernicus connection given a product id
}
