% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{parse_dataset_id}
\alias{parse_dataset_id}
\title{Parse one or more dataset_id values}
\usage{
parse_dataset_id(
  x = c("cmems_mod_glo_phy-cur_anfc_0.083deg_P1D-m",
    "cmems_mod_glo_phy-so_anfc_0.083deg_P1D-m",
    "cmems_mod_glo_phy-thetao_anfc_0.083deg_P1D-m",
    "cmems_mod_glo_phy-wcur_anfc_0.083deg_P1D-m",
    "cmems_mod_glo_phy_anfc_0.083deg-sst-anomaly_P1D-m",
    "cmems_mod_glo_phy_anfc_0.083deg_P1D-m"),
  detect_period = TRUE
)
}
\arguments{
\item{x}{one or more dataset_id values}

\item{detect_period}{logical, if TRUE, add a column, `.period` with 'day', 'month' etc
based upon a best guess from the `temporalres` segment (last segment of id)}
}
\value{
table of parsed elements
}
\description{
Parse one or more dataset_id values
}
