% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_body_system.R
\name{plot_system}
\alias{plot_system}
\title{Creates a static plot of the solar system}
\usage{
plot_system(df)
}
\arguments{
\item{variables}{}
}
\value{
a plot
}
\description{
Creates a static plot of the solar system
}
