% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{populate_script}
\alias{populate_script}
\title{Populate a script}
\usage{
populate_script(
  template = read_script(),
  bb = c(-77, -42.5, 36.5, 56.7),
  dates = Sys.Date() + c(-2, -1),
  times = c("01:00:00", "23:00:00"),
  depths = c(0.493, 0.4942),
  variables = c("bottomT", "mlotst", "siconc", "sithick", "so", "thetao", "uo", "usi",
    "vo", "vsi", "zos"),
  out_dir = ".",
  out_name = sprintf("copernicus_\%s.nc", format(Sys.Date(), "\%Y-\%m-\%d")),
  username = NULL,
  password = NULL,
  ...
)
}
\arguments{
\item{template}{the name of the script template}

\item{bb}{4 element bounding box [west, east, south, north]}

\item{dates}{2 element Date or Date-castable, [start, stop]}

\item{times}{2 element character vector, times (default is 12:00:00)}

\item{depths}{2 element numeric,  depth [min, max]}

\item{variables}{character vector, variable names}

\item{out_dir}{character, output directory}

\item{out_name}{character, output filename}

\item{username}{character or NULL, (if NULL get from options)}

\item{password}{character or NULL, (if NULL get from options)}

\item{...}{ignored}
}
\value{
2 element character vector of 'app' and 'param'
}
\description{
Populate a script
}
