% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{product_lut}
\alias{product_lut}
\title{Read a product LUT}
\usage{
product_lut(productid = "GLOBAL_ANALYSISFORECAST_PHY_001_024")
}
\arguments{
\item{productid}{cahr, the product indentifier}
}
\value{
tibble
}
\description{
Read a product LUT
}
