% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{read_dataset_catalog}
\alias{read_dataset_catalog}
\title{Read the json dataset catalog}
\usage{
read_dataset_catalog(
  filename = copernicus_path("catalogs/all_products_and_datasets.json"),
  tabulate = TRUE
)
}
\arguments{
\item{filename}{chr the path to the file}
}
\value{
a named list of json element
}
\description{
Read the json dataset catalog
}
