% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{read_dataset_description}
\alias{read_dataset_description}
\title{Read the description for a dataset}
\usage{
read_dataset_description(
  dataset_id = "cmems_mod_glo_phy_myint_0.083deg_P1D-m",
  tabulate = TRUE,
  flatten = TRUE,
  service_name = "arco-geo-series",
  path = copernicus_path("catalogs"),
  fetch = TRUE,
  x = NULL
)
}
\arguments{
\item{dataset_id}{chr, the dataset description to read}

\item{tabulate}{logical, if TRUE trasnform from list to nested table}

\item{flatten}{logical, if TRUE then flatten nested table (ignored if `tabulate = FALSE`)}

\item{service_name}{chr, if tabulating then the name of the service to tabulate. 
The default is used by the CLI.}

\item{path}{chr the copernicus data path}

\item{fetch}{logical, if TRUE and the data was not previously downloaded, then
try to fetch the data}

\item{x}{NULL or list, if not NULL then a list for a single dataset.  In this 
case no attempt is made to read the file.}
}
\value{
list, nested table or flat table
}
\description{
Read the description for a dataset
}
