% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{read_dataset_service}
\alias{read_dataset_service}
\title{Read a dataset service description as a table}
\usage{
read_dataset_service(x, flatten = TRUE)
}
\arguments{
\item{x}{a dataset node}

\item{flatten}{logical if TRUE then flatten otherwise return nested table}
}
\value{
table possibly nested or flattened
}
\description{
Read a dataset service description as a table
}
