% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{read_lut}
\alias{read_lut}
\title{Read a look-up-table (LUT)}
\usage{
read_lut(name = "global-analysis-forecast-phy-001-024")
}
\arguments{
\item{name}{character, the name of the LUT}
}
\value{
tibble of LUT
}
\description{
Read a look-up-table (LUT)
}
