% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{read_product_catalog}
\alias{read_product_catalog}
\title{Read the json product catalog}
\usage{
read_product_catalog(
  product_id = "all_products",
  path = copernicus_path("catalogs"),
  tabulate = TRUE,
  flatten = TRUE
)
}
\arguments{
\item{product_id}{chr, the prodict id to laod (or "all_products" if you have that)}

\item{path}{chr the path to the file}

\item{tabulate}{logical, if TRUE transform to a nested table}

\item{flatten}{logical, if TRUE and `tabulate` is TRUE then transofrm to a 
flat table}
}
\value{
a named list of json element
}
\description{
Read the json product catalog
}
