% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{read_region_lut}
\alias{read_region_lut}
\title{Read a regional LUT ala ("nwa_lut.csv")}
\usage{
read_region_lut(region = "nwa", path = copernicus_path("lut"))
}
\arguments{
\item{region}{char the name of the region}

\item{path}{char the path to the LUT}
}
\value{
tibble with "name", "longname" and "units"
}
\description{
Read a regional LUT ala ("nwa_lut.csv")
}
