% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{read_service_lut}
\alias{read_service_lut}
\title{Merge one or more LUTs together}
\usage{
read_service_lut(
  service = "global_analysisforecast_phy_001_024",
  path = copernicus_path("lut")
)
}
\arguments{
\item{service}{char the name of the service}

\item{path}{char, the path to the LUTs}
}
\value{
tibble or combined product luts for this service
  with "name", "longname" and "units"
}
\description{
Merge one or more LUTs together
}
