% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{read_stars_list}
\alias{read_stars_list}
\title{Read a ncdf object into a list of stars objects, possibly grouped by
dimensional dependence.}
\usage{
read_stars_list(filename, group = TRUE)
}
\arguments{
\item{filename}{str, the name of the file to read}

\item{group}{logical, if TRUE read variables grouped by dimensional dependence
if FALSE then each variable is read separately}
}
\value{
list of stars objects
}
\description{
Read a ncdf object into a list of stars objects, possibly grouped by
dimensional dependence.
}
