% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scan_file_copernicus.R
\name{scan_file_copernicus}
\alias{scan_file_copernicus}
\title{Scan COPERNICUS file(s) name}
\usage{
scan_file_copernicus(fnames)
}
\arguments{
\item{fnames}{\code{character} vector of file name(s)}
}
\value{
a \code{data.frame} with variables 'input' (input name(s)), 'product', 'date', 'tile', 'sensor' (VGT or PROBAV), 'version' (version of the algorithm)
}
\description{
Extract COPERNICUS product characteristics from file names
}
\details{
The geoland2-BioPar product follows the following naming  standard:
}
\examples{
fnames <- c('g2_BIOPAR_NDVI_201102030000_H19V4_VGT_V1.3.zip',
            'g2_BIOPAR_NDVI_201410010000_H19V4_PROBAV_V2.1.zip')
scan_file_copernicus(fnames)
}
\author{
Antoine Stevens
}
\references{
Baret et al. 2010. BioPar Product User Manual. Geoland2: Towards an Operational GMES Land Monitoring Core Service, 42 p.
}

