% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{select_database}
\alias{select_database}
\title{Select just the db columns}
\usage{
select_database(x, cols = database_variables())
}
\arguments{
\item{x}{database table}

\item{cols}{chr, the column names to keep}
}
\value{
a database table
}
\description{
Select just the db columns
}
