% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_copernicus_app}
\alias{set_copernicus_app}
\title{Write `copernicusmarine` app path to a file}
\usage{
set_copernicus_app(x = "copernicusmarine", filename = "~/.copernicusapp")
}
\arguments{
\item{x}{char, the path to the `copernicusmarine` app}

\item{filename}{the name the file to store the path as a single line of text}
}
\value{
NULL invisibly
}
\description{
Write `copernicusmarine` app path to a file
}
