% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_credentials}
\alias{set_credentials}
\title{Write copernicus credentials to a file}
\usage{
set_credentials(x = "username:password", filename = "~/.copernicuscredentials")
}
\arguments{
\item{x}{char, credentials as \code{"username:password"}}

\item{filename}{the name the file to store the credentials as a single line of text}
}
\value{
NULL invisibly
}
\description{
Write copernicus credentials to a file
}
