% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{set_point}
\alias{set_point}
\title{Set the "point" value for a stars object}
\usage{
set_point(x, point = NA)
}
\arguments{
\item{x}{stars object}

\item{point}{logical possible values are TRUE, FALSE and NA}
}
\value{
the input with x and y dim point values modified
}
\description{
Set the "point" value for a stars object
}
