% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{system_command}
\alias{system_command}
\title{Run either system() or system2() depending upon the host platform.}
\usage{
system_command(args, app = get_copernicus_app(), verbose = FALSE, ...)
}
\arguments{
\item{args}{chr a single character string of arguments for the app}

\item{app}{chr the path and name of the copernicusmarine app}

\item{verbose}{logical, if TRUE echo the command before issuing it}

\item{...}{arguments passed to either `system` or `system2`}
}
\value{
num, 0 if success, non-zero otherwise
}
\description{
We have encountered trouble using system2() on darwin/macos
}
