% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{tabulate_datasets}
\alias{tabulate_datasets}
\title{Tabulate dataset metadata}
\usage{
tabulate_datasets(x = read_dataset_catalog())
}
\arguments{
\item{x}{a datasets catalog node}
}
\value{
tibble
}
\description{
Tabulate dataset metadata
}
