% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{tabulate_product_catalog}
\alias{tabulate_product_catalog}
\title{Tabulate a product catalog}
\usage{
tabulate_product_catalog(x = read_product_catalog())
}
\arguments{
\item{x}{catalog list}
}
\value{
tibble
}
\description{
Tabulate a product catalog
}
