% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{time_as_string}
\alias{time_as_string}
\title{Convert a time to string}
\usage{
time_as_string(x)
}
\arguments{
\item{x}{char, Date or POSIXt class object}
}
\value{
the input as character
}
\description{
Convert a time to string
}
