% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{unpack_copernicus}
\alias{unpack_copernicus}
\title{Unpack a copernicus ncdf4 file}
\usage{
unpack_copernicus(filename, banded = FALSE)
}
\arguments{
\item{filename}{character, the full path specification}

\item{banded}{logical, see \code{\link{get_var}}.  Setting to FALSE
returns single band objects (depth and time dropped)}
}
\value{
list of \code{stars} objects
}
\description{
Unpack a copernicus ncdf4 file
}
