% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{unpack_dataset}
\alias{unpack_dataset}
\title{Unpack a single node into a tibble}
\usage{
unpack_dataset(x = get_dataset_node())
}
\arguments{
\item{x}{a single dataset catalog node}
}
\value{
a table of attribute data (short_name, standard_name, vars)
}
\description{
Unpack a single node into a tibble
}
