% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{unpack_product}
\alias{unpack_product}
\title{Unpack a product element}
\usage{
unpack_product(x)
}
\arguments{
\item{x}{json product list}
}
\value{
a nested product table
}
\description{
Unpack a product element
}
