% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_copernicus}
\alias{write_copernicus}
\title{Write a copernicus file given a stars object, database and path}
\usage{
write_copernicus(x, db, path = ".", check_path = TRUE)
}
\arguments{
\item{x}{a stars object}

\item{db}{database table}

\item{path}{chr, the output data path used by [compose_filename]}

\item{check_path}{logical, if TRUE then make sure the output path exists.  Note 
the final output path is generated by [compose_filename] and may not be the same
as `path`.}
}
\value{
the input stars object
}
\description{
Write a copernicus file given a stars object, database and path
}
