% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copernicus_options.R
\name{copernicus_options}
\alias{copernicus_options}
\title{Set or get options for the COPERNICUS package}
\usage{
copernicus_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for the COPERNICUS package
}
\note{
\code{resamplingType} is not working yet. Use function argument instead.
}
\section{Supported options}{

The following options are supported
\itemize{
   \item{\code{downloadPath}} {Path where the data should be downloaded}
   \item{\code{outPath}} {Path where the output of \code{get_copernicus} should be stored}
   \item{\code{outProj}} {\code{character} string in the PROJ.4 format (see \code{?\link[sp]{CRS}}). Default to '+init=epsg:32662', corresponding to the Plate Carr\'ee WGS 84}
   \item{\code{pixelSize}} {\code{character} string in the PROJ.4}
   \item{\code{resamplingType}} {Default resampling method. Default is 'near'. See \code{\link[gdalUtils]{gdalwarp}} for other options}
   \item{\code{gdalPath}} {Default is 'C:/OSGeo4W64/bin/' (this is when the OSGEO suite has been installed)}
   \item{\code{server}} {Default is 'http://land.copernicus.vgt.vito.be/PDF///datapool/Vegetation/'}
   \item{\code{user}} {user name to access COPERNICUS data portal}
   \item{\code{password}} {password associate with user name to access COPERNICUS data portal}
}
}
\examples{
copernicus_options('user' = 'nicolaus', password = 'copernicus')
copernicus_options()
}

