% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{create_lut}
\alias{create_lut}
\title{Generate a LUT suitable for the package}
\usage{
create_lut(
  x = "GLOBAL_ANALYSISFORECAST_BGC_001_028",
  prod = read_product_catalog(),
  save_lut = TRUE
)
}
\arguments{
\item{x}{chr the name of the product}

\item{prod}{table of products (unflattened)}

\item{save_lut}{log, if TRUE save to CSV format in `inst/lut`}
}
\value{
a table of look up values.  You'll edit this file to decide whihc to fetch
 and what depths to fetch from.
}
\description{
Generate a LUT suitable for the package
}
