% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{dataset_period}
\alias{dataset_period}
\title{Guess at the period of dataset given its id}
\usage{
dataset_period(
  x = c("cmems_mod_glo_phy_anfc_0.083deg_P1D-m", "cmems_mod_glo_phy_anfc_0.083deg_P1M-m")
)
}
\arguments{
\item{x}{one or more dataset_id values}
}
\value{
chr vector of "day", "month", NA (unknown) etc
}
\description{
Guess at the period of dataset given its id
}
