% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_copernicus}
\alias{fetch_copernicus}
\title{Retrieve a stars array of Copernicus data}
\usage{
fetch_copernicus(use = c("cli", "script")[1], ...)
}
\arguments{
\item{use}{char, download method one of 'cli' or 'script' (deprecated)}

\item{...}{
  Arguments passed on to \code{\link[=download_copernicus_cli]{download_copernicus_cli}}
  \describe{
    \item{\code{verbose}}{logical, if true pint the calling sequence excluding credentials}
    \item{\code{credentials}}{two element named character vector of \code{username} and \code{password}}
  }}
}
\description{
Retrieve a stars array of Copernicus data
}
