% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_copernicus}
\alias{fetch_copernicus}
\title{Retrieve a stars array of Copernicus data}
\usage{
fetch_copernicus(use = c("cli", "script")[1], bind = TRUE, ...)
}
\arguments{
\item{use}{char, download method one of 'cli' or 'script' (deprecated)}

\item{bind}{logical, if TRUE bind the elements of the returned into a single 
multi-attribute stars object.}

\item{...}{
  Arguments passed on to \code{\link[=download_copernicus_cli_subset]{download_copernicus_cli_subset}}
  \describe{
    \item{\code{verbose}}{logical, if true pint the calling sequence excluding credentials}
    \item{\code{stdout}}{logical or chr, if TRUE then return a charcater vector with output,
otherwise return numeric code.}
  }}
}
\description{
Retrieve a stars array of Copernicus data
}
