% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{fetch_copernicus_cli_subset}
\alias{fetch_copernicus_cli_subset}
\title{Fetch Copernicus data as a list of \code{stars} objects}
\usage{
fetch_copernicus_cli_subset(ofile = "output.nc", cleanup = TRUE, ...)
}
\arguments{
\item{ofile}{chr, the temporary (?) outfile}

\item{cleanup}{logical, if TRUE clean up files}

\item{...}{
  Arguments passed on to \code{\link[=download_copernicus_cli_subset]{download_copernicus_cli_subset}}
  \describe{
    \item{\code{verbose}}{logical, if true pint the calling sequence excluding credentials}
  }}
}
\value{
named list of stars objects (organized by variable)
}
\description{
This is a wrapper around \code{\link{download_copernicus_cli_subset}} that
hides the details and returns a list of \code{stars} objects.  The downloaded
file is deleted.
}
