% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{fetch_copernicus_cli_subset}
\alias{fetch_copernicus_cli_subset}
\title{Fetch Copernicus data as a list of \code{stars} objects}
\usage{
fetch_copernicus_cli_subset(
  ofile = copernicus_path("temp", paste0(dataset_id[1], ".nc")),
  cleanup = TRUE,
  ...
)
}
\arguments{
\item{cleanup}{logical, if TRUE clean up files}

\item{...}{
  Arguments passed on to \code{\link[=download_copernicus_cli_subset]{download_copernicus_cli_subset}}
  \describe{
    \item{\code{verbose}}{logical, if true pint the calling sequence excluding credentials}
  }}
}
\value{
stars object or NULL
}
\description{
This is a wrapper around \code{\link{download_copernicus_cli_subset}} that
hides the details and returns a \code{stars} object. Be aware that what gets
returned may have one or more degenerate dimensions (single element dimensions)
such as a single depth, time or both; it depends upon the request that you make.
}
