% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{flatten_product}
\alias{flatten_product}
\title{Flatten a particular product suite into one large table}
\usage{
flatten_product(
  x = dplyr::filter(read_product_catalog(), product_id ==
    "GLOBAL_ANALYSISFORECAST_PHY_001_024", grepl("P1D", dataset_id, fixed = TRUE))
)
}
\description{
Flatten a particular product suite into one large table
}
