% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{generate_filename}
\alias{generate_filename}
\title{Generate one or more file names for the data}
\usage{
generate_filename(
  x,
  id = "copernicus",
  depth = NULL,
  depth_signif = 4,
  period = guess_period(x),
  treatment = "raw",
  ext = "tif"
)
}
\arguments{
\item{x}{ncdf4 object or stars object}

\item{id}{character, the product or dataset identifier}

\item{depth, }{if NULL guess the depth, otherwise use the specified depth}

\item{depth_signif}{numeric, passed to \code{\link[base]{signif}} for rounding.}

\item{period}{char, the period description, guessed if not provided}

\item{treatment}{char the treatment description}

\item{ext}{char, the file extension of NULL to skip}
}
\description{
File names have the form of "id__datetime_depth_period_variable_treatment.ext"
Note the double underscore which permits parsing of the id if needed. Multiples
are ordered by first time, then variable and finally depth Presumably, period,
if known, is repeated
}
\details{
The pattern is datasetid__time_depth_period_var_treatment.ext
}
