% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_copernicus_app}
\alias{get_copernicus_app}
\title{Retrieve `copernicusmarine` app}
\usage{
get_copernicus_app(filename = "~/.copernicusapp", default = "copernicusmarine")
}
\arguments{
\item{filename}{the name the file}

\item{default}{chr the default to use if the file isn't found}
}
\value{
chr the known app (with path)
}
\description{
Retrieve `copernicusmarine` app
}
