% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{loc_index}
\alias{loc_index}
\title{Find the closest indices into a dimension for a given set of values}
\usage{
loc_index(x, value, what = "lon", make_rle = FALSE)
}
\arguments{
\item{x}{ncdf4 object}

\item{value}{vector of one of values (numeric for lon, lat, depth and POSIXct for time)}

\item{what}{character, the name of the dimension}

\item{make_rle}{logical, if TRUE use the first and last elements of value (which
may also be the first) to construct \code{[start, length]} encodings}
}
\value{
numeric vector if either indices closest to requested locations or
       two element \code{[start, length]} vector if \code{rle = TRUE}
}
\description{
This is useful to convert real dimensional values to array indices for making
data requests. The index assigned is always the dimensional step closest.
}
