% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{product_lut}
\alias{product_lut}
\title{Read a product LUT}
\usage{
product_lut(
  product_id = "GLOBAL_ANALYSISFORECAST_PHY_001_024",
  add_meta = TRUE
)
}
\arguments{
\item{product_id}{char, the product identifier}

\item{add_meta}{logical, if TRUE merge with `dataset_metadata`}
}
\value{
tibble
}
\description{
Read a product LUT
}
