% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{product_lut}
\alias{product_lut}
\title{Read a product LUT}
\usage{
product_lut(productid = "global-analysis-forecast-phy-001-024")
}
\arguments{
\item{productid}{char, the product identifier}
}
\value{
tibble
}
\description{
Read a product LUT
}
