% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_copernicus}
\alias{read_copernicus}
\title{Read one or more copernicus files}
\usage{
read_copernicus(db, path)
}
\arguments{
\item{db}{tibble, database of selected records}

\item{path}{char, the path to the data set}
}
\value{
stars object
}
\description{
By default the function tries to return an object with (variables of x, y, time) 
dimensions.  If multiple times are provided, then each var must be equally
represented.
}
