% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_copernicus}
\alias{read_copernicus}
\title{Read a copernicus ncdf4 file}
\usage{
read_copernicus(filename, banded = FALSE)
}
\arguments{
\item{filename}{character, the full path specification}

\item{banded}{logical, see \code{\link{get_var}}.  Setting to FALSE
returns single band objects (depth and time dropped)}
}
\value{
\code{stars} object
}
\description{
Read a copernicus ncdf4 file
}
