% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{read_product_catalog}
\alias{read_product_catalog}
\title{Read the json product catalog}
\usage{
read_product_catalog(
  filename = copernicus_path("catalogs/all_products.json"),
  tabulate = TRUE
)
}
\arguments{
\item{filename}{chr the path to the file}
}
\value{
a named list of json element
}
\description{
Read the json product catalog
}
