% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{read_script}
\alias{read_script}
\title{Read shell script into a characater array}
\usage{
read_script(
  filename = system.file("scripts/global-analysis-forecast-phy-001-024.sh", package =
    "copernicus")
)
}
\arguments{
\item{filename, }{character, the name of the script file}
}
\value{
named character vector with 'app' and 'param'
}
\description{
Read shell script into a characater array
}
